Tidy up some more warnings generated when using --enable-warnings.
Legacy-ID: 310

+++ spec128.c
@@ -117,7 +117,7 @@ DWORD spec128_contend_port( WORD port )
 
 static DWORD spec128_contend_delay( void )
 {
-  DWORD tstates_through_line;
+  WORD tstates_through_line;
   
   /* No contention in the upper border */
   if( tstates < machine_current->line_times[ DISPLAY_BORDER_HEIGHT ] )
