+++ specplus3.c
@@ -191,7 +191,7 @@ DWORD specplus3_contend_port( WORD port )
 
 static DWORD specplus3_contend_delay( void )
 {
-  DWORD tstates_through_line;
+  WORD tstates_through_line;
   
   /* No contention in the upper border */
   if( tstates < machine_current->line_times[ DISPLAY_BORDER_HEIGHT ] )
